project refactor into packages
