+++ app/src/androidTest/java/com/sdpteam/connectout/authentication/ExampleInstrumentedTest.java
@@ -2,13 +2,12 @@ package com.sdpteam.connectout.authentication;
 
 import static org.junit.Assert.assertEquals;
 
-import android.content.Context;
-
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.platform.app.InstrumentationRegistry;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import android.content.Context;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
 
 /**
  * Instrumented test, which will execute on an Android device.
