+++ app/src/androidTest/java/com/sdpteam/connectout/map/MapActivityTest.java
@@ -1,4 +1,4 @@
-package com.sdpteam.connectout;
+package com.sdpteam.connectout.map;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
@@ -6,14 +6,16 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
-import androidx.test.espresso.ViewInteraction;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import com.sdpteam.connectout.R;
+
+import androidx.test.espresso.ViewInteraction;
+import androidx.test.espresso.matcher.ViewMatchers;
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 @RunWith(AndroidJUnit4.class)
 public class MapActivityTest {
@@ -23,7 +25,7 @@ public class MapActivityTest {
 
     @Test
     public void testMapDisplayed() {
-        ViewInteraction mapFragment = onView(withId(R.id.map)).check(matches(isDisplayed()));
+        ViewInteraction mapFragment = onView(ViewMatchers.withId(R.id.map)).check(matches(isDisplayed()));
         onView(withId(R.id.refresh_button)).perform(click());
         onView(withId(R.id.map)).check(matches(isDisplayed()));
     }
