+++ app/src/androidTest/java/com/sdpteam/connectout/utils/LiveDataTestUtil.java
@@ -1,4 +1,4 @@
-package com.sdpteam.connectout;
+package com.sdpteam.connectout.utils;
 
 import java.util.concurrent.CompletableFuture;
