--- app/src/main/java/com/sdpteam/connectout/GreetingActivity.java
@@ -1,21 +1,0 @@
-package com.sdpteam.connectout;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.widget.TextView;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class GreetingActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_greeting);
-
-        Intent intent = getIntent();
-        String value = intent.getStringExtra("key");
-
-        TextView t = findViewById(R.id.greetingMessage);
-        t.setText(value);
-    }
-}
