+++ app/src/main/java/com/sdpteam/connectout/MainActivity.java
@@ -1,8 +1,9 @@
 package com.sdpteam.connectout;
 
+import com.sdpteam.connectout.drawer.DrawerActivity;
+
 import android.content.Intent;
 import android.os.Bundle;
-
 import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
@@ -11,17 +12,7 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-/*
-        EditText t = findViewById(R.id.mainName);
-        Button b = findViewById(R.id.mainGoButton);
-        /*
-        b.setOnClickListener(v -> {
-            Intent myIntent = new Intent(getApplicationContext(), GreetingActivity.class);
-            String msg = "Hello " + t.getText().toString() + "!";
-            myIntent.putExtra("key", msg);
-            MainActivity.this.startActivity(myIntent);
-        });*/
-
+        // Don't put anything here, just choose which activity to redirect to
         Intent drawerIntent = new Intent(getApplicationContext(), DrawerActivity.class);
         this.startActivity(drawerIntent);
     }
