+++ app/src/main/java/com/sdpteam/connectout/drawer/DrawerActivity.java
@@ -1,9 +1,11 @@
-package com.sdpteam.connectout;
+package com.sdpteam.connectout.drawer;
+
+import com.google.android.material.navigation.NavigationView;
+import com.sdpteam.connectout.R;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
-
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
@@ -12,13 +14,6 @@ import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-
-import com.google.android.material.navigation.NavigationView;
-import com.sdpteam.connectout.fragments.FilterFragment;
-import com.sdpteam.connectout.fragments.HomeFragment;
-import com.sdpteam.connectout.fragments.MyAccountFragment;
-import com.sdpteam.connectout.fragments.MyEventsFragment;
-
 
 public class DrawerActivity extends AppCompatActivity {
     @SuppressLint("NonConstantResourceId")
@@ -55,7 +50,6 @@ public class DrawerActivity extends AppCompatActivity {
             displayFragment(item.getItemId());
             return true;
         });
-
     }
 
     /**
@@ -87,30 +81,23 @@ public class DrawerActivity extends AppCompatActivity {
         if (itemId == R.id.menu_home) {
             //Go back to the Map/List view
             replaceFragment(new HomeFragment(), R.id.drawer_fragment_container);
-
         }
         if (itemId == R.id.menu_my_account) {
             //Go check out your account
             replaceFragment(new MyAccountFragment(), R.id.drawer_fragment_container);
-
         }
         if (itemId == R.id.menu_my_events) {
             //Go check out your events
             replaceFragment(new MyEventsFragment(), R.id.drawer_fragment_container);
-
         }
         if (itemId == R.id.menu_filters) {
             //Add add to the Map/List view, a TextEdit in order to filter what is displayed
             FilterFragment filterFragment = new FilterFragment();
             replaceFragment(filterFragment, R.id.drawer_fragment_container);
-
         }
         if (itemId == R.id.menu_logout) {
             Intent logOutIntent = new Intent(getApplicationContext(), LogInActivity.class);
             startActivity(logOutIntent);
-
         }
     }
-
-
 }
