+++ app/src/main/java/com/sdpteam/connectout/drawer/FilterFragment.java
@@ -1,18 +1,15 @@
-package com.sdpteam.connectout.fragments;
+package com.sdpteam.connectout.drawer;
+
+import com.sdpteam.connectout.R;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.sdpteam.connectout.R;
-
-
 public class FilterFragment extends Fragment {
-
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
@@ -21,5 +18,4 @@ public class FilterFragment extends Fragment {
         this.getChildFragmentManager().beginTransaction().replace(R.id.filter_container, new HomeFragment()).commit();
         return inflater.inflate(R.layout.fragment_filter, container, false);
     }
-
 }
