+++ app/src/main/java/com/sdpteam/connectout/drawer/LogInActivity.java
@@ -1,7 +1,8 @@
-package com.sdpteam.connectout;
+package com.sdpteam.connectout.drawer;
+
+import com.sdpteam.connectout.R;
 
 import android.os.Bundle;
-
 import androidx.appcompat.app.AppCompatActivity;
 
 public class LogInActivity extends AppCompatActivity {
