+++ app/src/main/java/com/sdpteam/connectout/drawer/MyAccountFragment.java
@@ -1,19 +1,19 @@
-package com.sdpteam.connectout.fragments;
+package com.sdpteam.connectout.drawer;
+
+import com.sdpteam.connectout.R;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.sdpteam.connectout.R;
-
-public class HomeFragment extends Fragment {
+public class MyAccountFragment extends Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_home, container, false);
+
+        return inflater.inflate(R.layout.fragment_my_account, container, false);
     }
 }
