+++ app/src/main/res/layout/activity_log_in.xml
@@ -1,70 +1,70 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:viewBindingIgnore="true"
-    tools:context=".LogInActivity">
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:viewBindingIgnore="true"
+        tools:context=".drawer.LogInActivity">
 
     <TextView
-        android:id="@+id/login_app_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_marginVertical="10dp"
-        android:text="@string/app_name"
-        android:textSize="50sp"
-        android:layout_below="@id/login_app_image" />
+            android:id="@+id/login_app_name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:layout_marginVertical="10dp"
+            android:text="@string/app_name"
+            android:textSize="50sp"
+            android:layout_below="@id/login_app_image" />
 
     <ImageView
-        android:layout_marginTop="15dp"
-        android:id="@+id/login_app_image"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        tools:srcCompat="@tools:sample/avatars"
-        android:contentDescription="@string/connect_out_logo" />
+            android:layout_marginTop="15dp"
+            android:id="@+id/login_app_image"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            tools:srcCompat="@tools:sample/avatars"
+            android:contentDescription="@string/connect_out_logo" />
 
     <EditText
-        android:id="@+id/login_person_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@id/login_app_name"
-        android:layout_marginVertical="10dp"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:hint="@string/login_name"
-        android:autofillHints="Username" />
+            android:id="@+id/login_person_name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:layout_below="@id/login_app_name"
+            android:layout_marginVertical="10dp"
+            android:ems="10"
+            android:inputType="textPersonName"
+            android:hint="@string/login_name"
+            android:autofillHints="Username" />
 
     <EditText
-        android:id="@+id/login_password"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/login_person_name"
-        android:layout_marginVertical="10dp"
-        android:layout_centerHorizontal="true"
-        android:ems="10"
-        android:inputType="textPassword"
-        android:hint="@string/login_password"
-        android:autofillHints="Password" />
+            android:id="@+id/login_password"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/login_person_name"
+            android:layout_marginVertical="10dp"
+            android:layout_centerHorizontal="true"
+            android:ems="10"
+            android:inputType="textPassword"
+            android:hint="@string/login_password"
+            android:autofillHints="Password" />
 
     <Button
-        android:id="@+id/login_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@id/login_password"
-        android:layout_marginVertical="10dp"
-        android:text="@string/log_in_button" />
+            android:id="@+id/login_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:layout_below="@id/login_password"
+            android:layout_marginVertical="10dp"
+            android:text="@string/log_in_button" />
 
     <Button
-        android:id="@+id/sign_in_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@id/login_button"
-        android:layout_marginVertical="10dp"
-        android:text="@string/sign_in_button" />
+            android:id="@+id/sign_in_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:layout_below="@id/login_button"
+            android:layout_marginVertical="10dp"
+            android:text="@string/sign_in_button" />
 
 </RelativeLayout>
