+++ app/src/main/res/layout/fragment_my_account.xml
@@ -1,15 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/my_account_fragment"
-    tools:viewBindingIgnore="true">
-
-    <TextView
-        android:id="@+id/text_my_account"
+        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        />
+        android:id="@+id/my_account_fragment"
+        tools:viewBindingIgnore="true">
+
+    <TextView
+            android:id="@+id/text_my_account"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
 
 </FrameLayout>
