+++ app/src/main/res/menu/activity_menu_drawer.xml
@@ -1,30 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
+        xmlns:tools="http://schemas.android.com/tools"
+        tools:showIn="navigation_view">
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/menu_home"
-            android:icon="@drawable/home_image"
-            android:title="@string/menu_home" />
+                android:id="@+id/menu_home"
+                android:icon="@drawable/home_image"
+                android:title="@string/menu_home" />
         <item
-            android:id="@+id/menu_filters"
-            android:icon="@drawable/filter_image"
-            android:title="@string/menu_filters" />
+                android:id="@+id/menu_filters"
+                android:icon="@drawable/filter_image"
+                android:title="@string/menu_filters" />
         <item
-            android:id="@+id/menu_my_account"
-            android:icon="@drawable/account_image"
-            android:title="@string/menu_my_account" />
+                android:id="@+id/menu_my_account"
+                android:icon="@drawable/account_image"
+                android:title="@string/menu_my_account" />
         <item
-            android:id="@+id/menu_my_events"
-            android:icon="@drawable/event_image"
-            android:title="@string/menu_my_events" />
+                android:id="@+id/menu_my_events"
+                android:icon="@drawable/event_image"
+                android:title="@string/menu_my_events" />
     </group>
 
     <item
-        android:id="@+id/menu_logout"
-        android:icon="@drawable/logout_image"
-        android:title="@string/menu_log_out" />
+            android:id="@+id/menu_logout"
+            android:icon="@drawable/logout_image"
+            android:title="@string/menu_log_out" />
 
 </menu>
