+++ app/src/test/java/com/sdpteam/connectout/EventTest.java
@@ -2,9 +2,10 @@ package com.sdpteam.connectout;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
+
+import com.sdpteam.connectout.map.Event;
 
 public class EventTest {
     @Test
